#### PR description:

This PR speeds up the E/gamma HLT pixel matching. It uses work done by Calib Fangmeier to allow the pixel matching to run on triplet cleaned doublets which at a small efficiency cost will greatly increase speed (the path needs to enable this so its of for now). 

Additionally it does some optimization where expensive variables are now computed outside a loop, it also uses parametric b field for speed where appropriate and finally adds a minimum et cut for superclusters to pixel match. 

This benefits Run2/Run3 configs but the real reason is for PhaseII 200PU where the e/gamma hlt pixel matching was taking ~30mins an event. It now takes ~2seconds and could be further optimised.

No / very small changes are expected for existing HLT workflows although the timing is reduced. 

I validated this on an electron gun sample for 2021 MC conditions (the samples used for the Run3 projections) using /dev/CMSSW_11_0_0/GRun/V10

![image](https://user-images.githubusercontent.com/5021403/68960414-6bb77a80-07d0-11ea-8e2d-15c676af96a0.png)

Timing:
TimeReport   0.015958     0.032457     0.032457  hltEgammaElectronPixelSeeds
goes to
TimeReport   0.007012     0.014263     0.014263  hltEgammaElectronPixelSeeds

Doing a full timing study  on 2018 data I see a drop from 240ms to 236ms with the drops happening in electron paths as expected.