@cmsbuild 2019-11-15
A new Pull Request was created by  @Sam-Harper (Sam Harper) for master.

It involves the following packages:

RecoEgamma/EgammaElectronAlgos
RecoEgamma/EgammaElectronProducers
RecoEgamma/EgammaHLTProducers


@perrotta, @Martin-Grunewald, @cmsbuild, @fwyzard, @slava77 can you please review it and eventually sign? Thanks.
@battibass, @jainshilpi, @rovere, @lgray, @calderona, @HuguesBrun, @Fedespring, @afiqaize, @cericeci, @varuns23, @folguera, @sobhatta this is something you requested to watch as well.
@davidlange6, @slava77, @fabiocos you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>