@Sam-Harper 2019-11-18
ah I forgot to include Martin as this is really a HLT module
 @Martin-Grunewald 

At somepoint RECO will use this though, but not for 11_0_0