@perrotta 2019-11-19
@Sam-Harper , just to verify that I understood it correctly:
- In single electron gun the time taken by pixel matching goes down from 32 ms to 14 ms with this PR
- In PU200 Phase2 samples the time taken by pixel matching goes down from ~1800 s to ~2 s with this PR ("and it could be further optimised")

Do you confirm that I got it correctly and that there is such an impressive speed-up with PU200?