@Sam-Harper 2019-11-19
> Do you confirm that I got it correctly and that there is such an impressive speed-up with PU200?

Yes, phaseII e/gamma pixel matching was taking ~30mins to run per event at 200PU. Its not really an impressive speed up, we knew this module was doing things suboptimally just there were more important things to fix. As it combines pixel pairs, it scales extremely poorly with number of pixel hits.   Once it became a blocker for PhaseII, it got fixed. 

It was a combination of code optimizations and config changes for PhaseII, the Run2 stays with the old config and just gets the code optimizations. 

The 15ms->7ms was actually measured on data using a proper timing setup. 

This is something that would be nice to have in for 11_0_0 but then again it the 11_0_0 production will not use this as only a fake HLT menu will be run for the Phase II production.