@perrotta 2019-11-20
+1
- This code is only run at HLT, by now, and not in offline reco
- From a visual inspection, its implementation corresponds to what declared in the PR description
- The validation provided shows the improvement in the cpu consumption of the pixel matcher, and no effect on the inspected HLT results
- Jenkins tests show no differences in reco outputs (of course), and at a first glance not even in the HLT outputs